Created a task that runs on a minute interval that will work on cleaning up all zombie workers.
