+++ bld_commands/src/server/command.rs
@@ -1,18 +1,8 @@
 use crate::BldCommand;
-use actix_web::{middleware, rt::System, web, App, HttpServer};
+use actix_web::rt::System;
 use bld_config::{definitions::VERSION, BldConfig};
-use bld_core::database::new_connection_pool;
-use bld_core::high_avail::HighAvail;
-use bld_core::proxies::ServerPipelineProxy;
-use bld_server::endpoints::{
-    auth_redirect, deps, ha_append_entries, ha_install_snapshot, ha_vote, hist, home, inspect,
-    list, pull, push, remove, stop,
-};
-use bld_server::sockets::{ws_exec, ws_high_avail, ws_monit};
 use clap::{App as ClapApp, Arg, ArgMatches, SubCommand};
-use std::env::set_var;
-use std::sync::Arc;
-use tracing::{debug, info};
+use tracing::debug;
 
 static SERVER: &str = "server";
 static HOST: &str = "host";
@@ -23,55 +13,6 @@ pub struct ServerCommand;
 impl ServerCommand {
     pub fn boxed() -> Box<dyn BldCommand> {
         Box::new(Self)
-    }
-
-    async fn start(config: BldConfig, host: &str, port: i64) -> anyhow::Result<()> {
-        info!("starting bld server at {}:{}", host, port);
-        let pool = new_connection_pool(&config.local.db)?;
-        let ha = web::Data::new(HighAvail::new(&config, pool.clone()).await?);
-        let pool = web::Data::new(pool);
-        let cfg = web::Data::new(config);
-        let prx = web::Data::new(ServerPipelineProxy::new(
-            Arc::clone(&cfg),
-            Arc::clone(&pool),
-        ));
-        set_var("RUST_LOG", "actix_server=info,actix_web=debug");
-        HttpServer::new(move || {
-            App::new()
-                .app_data(cfg.clone())
-                .app_data(ha.clone())
-                .app_data(pool.clone())
-                .app_data(prx.clone())
-                .wrap(middleware::Logger::default())
-                .service(ha_append_entries)
-                .service(ha_install_snapshot)
-                .service(ha_vote)
-                .service(home)
-                .service(auth_redirect)
-                .service(hist)
-                .service(list)
-                .service(remove)
-                .service(push)
-                .service(deps)
-                .service(pull)
-                .service(stop)
-                .service(inspect)
-                .service(web::resource("/ws-exec/").route(web::get().to(ws_exec)))
-                .service(web::resource("/ws-monit/").route(web::get().to(ws_monit)))
-                .service(web::resource("/ws-ha/").route(web::get().to(ws_high_avail)))
-        })
-        .bind(format!("{host}:{port}"))?
-        .run()
-        .await?;
-        Ok(())
-    }
-
-    pub fn spawn(config: BldConfig, host: String, port: i64) -> anyhow::Result<()> {
-        debug!("starting actix system");
-        System::new().block_on(async move {
-            let _ = Self::start(config, &host, port).await;
-        });
-        Ok(())
     }
 }
 
@@ -109,7 +50,9 @@ impl BldCommand for ServerCommand {
             .map(|port| port.parse::<i64>().unwrap_or(config.local.port))
             .unwrap_or(config.local.port);
         debug!("running {SERVER} subcommand with --host: {host} --port: {port}",);
-        Self::spawn(config, host, port)?;
+        System::new().block_on(async move {
+            let _ = bld_server::start(config, &host, port).await;
+        });
         Ok(())
     }
 }
