+++ bld_server/src/data/mod.rs
@@ -1,0 +1,3 @@
+mod worker;
+
+pub use worker::*;
